requirements.txt
apscheduler
python-binance
termcolor
python-dotenv
openai
google-generativeai